Extract API transport helpers
