+++ service-worker.js
@@ -80,6 +80,7 @@ const APP_SHELL = [
   '/js/utils.js',
   '/js/theme.js',
   '/js/api.js',
+  '/js/api-transport.js',
   '/js/startup-foundation.js',
   '/js/startup-profile.js',
   '/js/startup-oauth-callbacks.js',
